split access rights to own module
