+++ user-access-internal-model/src/test/java/no/unit/nva/useraccessmanagement/dao/EntityUtils.java
@@ -1,6 +1,6 @@
 package no.unit.nva.useraccessmanagement.dao;
 
-import static no.unit.nva.useraccessmanagement.dao.AccessRight.APPROVE_DOI_REQUEST;
+import static no.unit.useraccessserivce.accessrights.AccessRight.APPROVE_DOI_REQUEST;
 import static nva.commons.utils.JsonUtils.objectMapper;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.io.InputStream;
@@ -13,6 +13,7 @@ import no.unit.nva.testutils.HandlerRequestBuilder;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidEntryInternalException;
 import no.unit.nva.useraccessmanagement.model.RoleDto;
 import no.unit.nva.useraccessmanagement.model.UserDto;
+import no.unit.useraccessserivce.accessrights.AccessRight;
 
 public final class EntityUtils {
