+++ user-access-rights/build.gradle
@@ -1,0 +1,39 @@
+dependencies {
+
+    compileOnly group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: project.ext.jacksonVersion
+    testImplementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: project.ext.jacksonVersion
+
+    compileOnly group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: project.ext.jacksonVersion
+    testImplementation group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: project.ext.jacksonVersion
+
+    testImplementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jdk8', version: project.ext.jacksonVersion
+    testImplementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jsr310', version: project.ext.jacksonVersion
+
+    testImplementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: project.ext.jacksonVersion
+
+    compileOnly project(":user-access-public-model")
+    testImplementation project(":user-access-public-model")
+
+    compileOnly project(":user-access-commons")
+    testImplementation project(":user-access-commons")
+
+    compileOnly project(":user-access-errors")
+    testImplementation project(":user-access-errors")
+
+    testImplementation (group: 'com.github.BIBSYSDEV', name: 'nva-commons', version: project.ext.nvaCommonsVersion) {
+        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-core'
+        exclude group: 'com.fasterxml.jackson.datatype', module: 'jackson-datatype-jdk8'
+        exclude group: 'com.fasterxml.jackson.datatype', module: 'jackson-datatype-jsr310'
+        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-databind'
+        exclude group: 'com.fasterxml.jackson.module', module: 'jackson-module-parameter-names'
+        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-annotations'
+        exclude group: 'com.amazonaws', module: 'aws-java-sdk-sts'
+        exclude group: 'com.amazonaws', module: 'aws-java-sdk-secretsmanager'
+        exclude group: 'com.amazonaws', module: 'aws-lambda-java-events'
+        exclude group: 'com.amazonaws', module: 'aws-java-sdk-lambda'
+    }
+
+
+
+
+}
