+++ user-access-rights/src/test/java/no/unit/useraccessserivce/accessrights/AccessRightTest.java
@@ -1,4 +1,4 @@
-package no.unit.nva.useraccessmanagement.dao;
+package no.unit.useraccessserivce.accessrights;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
@@ -7,7 +7,6 @@ import static org.hamcrest.core.StringContains.containsString;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.io.IOException;
-import no.unit.nva.useraccessmanagement.exceptions.InvalidAccessRightException;
 import nva.commons.utils.JsonUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
