+++ user-access-rights/src/test/java/no/unit/useraccessserivce/accessrights/InvalidAccessRightExceptionTest.java
@@ -1,7 +1,7 @@
-package no.unit.nva.useraccessmanagement.exceptions;
+package no.unit.useraccessserivce.accessrights;
 
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.StringContains.containsString;
+import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.Test;
 
 public class InvalidAccessRightExceptionTest {
@@ -11,6 +11,6 @@ public class InvalidAccessRightExceptionTest {
     @Test
     public void invalidAccessRightExceptionMessageContainsTheInvalidAccessRight() {
         InvalidAccessRightException exception = new InvalidAccessRightException(INVALID_ACCESS_RIGHT);
-        assertThat(exception.getMessage(), containsString(INVALID_ACCESS_RIGHT));
+        MatcherAssert.assertThat(exception.getMessage(), containsString(INVALID_ACCESS_RIGHT));
     }
 }
