+++ user-access-service/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     api project(":user-access-commons")
     api project(":user-access-errors")
     api project(":user-access-public-model")
+    api project(":user-access-rights")
 
 
     api(group: 'com.amazonaws', name: 'aws-java-sdk-dynamodb', version: project.ext.awsSdkVersion) {
