+++ user-access-service/src/test/java/no/unit/nva/database/DatabaseServiceTest.java
@@ -30,13 +30,13 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import no.unit.nva.useraccessmanagement.constants.DatabaseIndexDetails;
-import no.unit.nva.useraccessmanagement.dao.AccessRight;
 import no.unit.nva.useraccessmanagement.dao.RoleDb;
 import no.unit.nva.useraccessmanagement.dao.UserDb;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidEntryInternalException;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidInputException;
 import no.unit.nva.useraccessmanagement.model.RoleDto;
 import no.unit.nva.useraccessmanagement.model.UserDto;
+import no.unit.useraccessserivce.accessrights.AccessRight;
 import nva.commons.exceptions.commonexceptions.ConflictException;
 import nva.commons.exceptions.commonexceptions.NotFoundException;
 import nva.commons.utils.SingletonCollector;
