+++ user-access-service/src/test/java/no/unit/nva/database/EntityUtils.java
@@ -5,10 +5,10 @@ import java.lang.reflect.Method;
 import java.util.Collections;
 import java.util.Set;
 import java.util.stream.Collectors;
-import no.unit.nva.useraccessmanagement.dao.AccessRight;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidEntryInternalException;
 import no.unit.nva.useraccessmanagement.model.RoleDto;
 import no.unit.nva.useraccessmanagement.model.UserDto;
+import no.unit.useraccessserivce.accessrights.AccessRight;
 
 public final class EntityUtils {
